ml-data/requirements.txt
pyserial>=3.5
pandas>=1.5.0
numpy>=1.23.0
tensorflow>=2.12.0
scikit-learn>=1.2.0
